refactor(gui): extract main-window layout into window.rs
First cut at splitting the 4630-line lib.rs. Window geometry is the most
self-contained piece: pure arithmetic over PhysicalRect/PhysicalWindowSize plus
one function that touches a real window, and it already had three tests.

Moves the geometry helpers, the three Physical* structs, the two DEFAULT_
constants and the tests into window.rs; lib.rs keeps only the
reset_main_window_layout command. No behavior change -- same 74 tests pass, and
LogicalSize/PhysicalPosition drop out of the lib.rs import list.

lib.rs 4630 -> 4435.
