+++ src-tauri/src/lib.rs
@@ -1,4 +1,5 @@
 mod batch_auth;
+mod window;
 
 use std::collections::HashMap;
 use std::sync::atomic::{AtomicBool, Ordering};
@@ -9,7 +10,7 @@ use std::thread::JoinHandle;
 use std::time::Duration;
 
 use serde::{Deserialize, Serialize};
-use tauri::{AppHandle, Emitter, LogicalSize, Manager, PhysicalPosition, RunEvent, State};
+use tauri::{AppHandle, Emitter, Manager, RunEvent, State};
 use tauri_plugin_log::{RotationStrategy, Target, TargetKind, TimezoneStrategy};
 use tyutool_core::{
     serial_debug_fail_backfill_if_current, serial_debug_finish_backfill_if_current,
@@ -124,8 +125,6 @@ struct SerialDebugFilterAddArgs {
     color: String,
 }
 
-const DEFAULT_MAIN_WINDOW_WIDTH: f64 = 1280.0;
-const DEFAULT_MAIN_WINDOW_HEIGHT: f64 = 800.0;
 const SERIAL_DEBUG_CHUNK_FLUSH_MS: u64 = 12;
 const SERIAL_DEBUG_CHUNK_FLUSH_BYTES: usize = 32 * 1024;
 const SERIAL_DEBUG_CHUNK_QUEUE_CAPACITY: usize = 256;
@@ -352,26 +351,6 @@ fn spawn_serial_debug_chunk_bridge(
         }
     });
     handle
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-struct PhysicalRect {
-    x: i32,
-    y: i32,
-    width: u32,
-    height: u32,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-struct PhysicalWindowSize {
-    width: u32,
-    height: u32,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-struct PhysicalWindowPosition {
-    x: i32,
-    y: i32,
 }
 
 /// Detect the installation type at runtime based on the executable's path.
@@ -2253,112 +2232,9 @@ fn set_log_level(level: String) -> Result<(), String> {
 /// applied on purpose: the window's `minWidth`/`minHeight` in tauri.conf.json
 /// already enforces one, and forcing a minimum here would push the window off a
 /// work area smaller than that minimum.
-fn fit_logical_dimension(default: f64, available: f64) -> f64 {
-    if !available.is_finite() || available <= 0.0 {
-        return default;
-    }
-    default.min(available)
-}
-
-fn default_main_window_logical_size(
-    work_area: PhysicalRect,
-    scale_factor: f64,
-) -> LogicalSize<f64> {
-    let scale_factor = if scale_factor.is_finite() && scale_factor > 0.0 {
-        scale_factor
-    } else {
-        1.0
-    };
-    let available_width = f64::from(work_area.width) / scale_factor;
-    let available_height = f64::from(work_area.height) / scale_factor;
-
-    LogicalSize::new(
-        fit_logical_dimension(DEFAULT_MAIN_WINDOW_WIDTH, available_width),
-        fit_logical_dimension(DEFAULT_MAIN_WINDOW_HEIGHT, available_height),
-    )
-}
-
-fn clamp_axis(position: i32, size: u32, work_start: i32, work_extent: u32) -> i32 {
-    if size >= work_extent {
-        return work_start;
-    }
-
-    let min = i64::from(work_start);
-    let max = min + i64::from(work_extent) - i64::from(size);
-    i64::from(position).clamp(min, max) as i32
-}
-
-fn clamp_outer_position_to_work_area(
-    x: i32,
-    y: i32,
-    outer_size: PhysicalWindowSize,
-    work_area: PhysicalRect,
-) -> PhysicalWindowPosition {
-    PhysicalWindowPosition {
-        x: clamp_axis(x, outer_size.width, work_area.x, work_area.width),
-        y: clamp_axis(y, outer_size.height, work_area.y, work_area.height),
-    }
-}
-
-fn physical_rect_from_tauri(rect: &tauri::PhysicalRect<i32, u32>) -> PhysicalRect {
-    PhysicalRect {
-        x: rect.position.x,
-        y: rect.position.y,
-        width: rect.size.width,
-        height: rect.size.height,
-    }
-}
-
-/// Default main window size + safe visible placement (matches `tauri.conf.json` when it fits).
-fn apply_default_main_window_layout(app: &AppHandle) -> Result<(), String> {
-    if let Some(win) = app.get_webview_window("main") {
-        let monitor = win
-            .current_monitor()
-            .map_err(|e| e.to_string())?
-            .or_else(|| win.primary_monitor().ok().flatten())
-            .or_else(|| {
-                win.available_monitors()
-                    .ok()
-                    .and_then(|monitors| monitors.into_iter().next())
-            });
-
-        if let Some(monitor) = monitor {
-            let work_area = physical_rect_from_tauri(monitor.work_area());
-            let size = default_main_window_logical_size(work_area, monitor.scale_factor());
-            win.set_size(size).map_err(|e| e.to_string())?;
-            win.center().map_err(|e| e.to_string())?;
-
-            let outer_position = win.outer_position().map_err(|e| e.to_string())?;
-            let outer_size = win.outer_size().map_err(|e| e.to_string())?;
-            let clamped = clamp_outer_position_to_work_area(
-                outer_position.x,
-                outer_position.y,
-                PhysicalWindowSize {
-                    width: outer_size.width,
-                    height: outer_size.height,
-                },
-                work_area,
-            );
-
-            if clamped.x != outer_position.x || clamped.y != outer_position.y {
-                win.set_position(PhysicalPosition::new(clamped.x, clamped.y))
-                    .map_err(|e| e.to_string())?;
-            }
-        } else {
-            win.set_size(LogicalSize::new(
-                DEFAULT_MAIN_WINDOW_WIDTH,
-                DEFAULT_MAIN_WINDOW_HEIGHT,
-            ))
-            .map_err(|e| e.to_string())?;
-            win.center().map_err(|e| e.to_string())?;
-        }
-    }
-    Ok(())
-}
-
 #[tauri::command]
 fn reset_main_window_layout(app: AppHandle) -> Result<(), String> {
-    apply_default_main_window_layout(&app)
+    window::apply_default_main_window_layout(&app)
 }
 
 /// Registry of filesystem paths the renderer may write to.
@@ -3718,7 +3594,7 @@ pub fn run() {
                 RunEvent::Ready => {
                     // After the event loop is ready: layout, then show (window starts `visible: false`
                     // so the compositor / session restore does not paint a wrong geometry first).
-                    let _ = apply_default_main_window_layout(app_handle);
+                    let _ = window::apply_default_main_window_layout(app_handle);
                     if let Some(win) = app_handle.get_webview_window("main") {
                         let _ = win.show();
                     }
@@ -3728,7 +3604,7 @@ pub fn run() {
                         std::thread::sleep(Duration::from_millis(280));
                         let h2 = h.clone();
                         let _ = h.run_on_main_thread(move || {
-                            let _ = apply_default_main_window_layout(&h2);
+                            let _ = window::apply_default_main_window_layout(&h2);
                         });
                     });
                 }
@@ -3788,60 +3664,6 @@ mod tests {
     fn serial_debug_device_reset_session_requires_open_session() {
         let err = serial_debug_device_reset_session(None, "T5AI").unwrap_err();
         assert_eq!(err, "serial debug not open");
-    }
-
-    #[test]
-    fn default_layout_shrinks_to_fit_high_dpi_work_area() {
-        let work_area = PhysicalRect {
-            x: 0,
-            y: 0,
-            width: 1920,
-            height: 1040,
-        };
-
-        let size = default_main_window_logical_size(work_area, 1.5);
-
-        assert_eq!(size.width, 1280.0);
-        assert!(size.height < DEFAULT_MAIN_WINDOW_HEIGHT);
-        assert!(size.height <= 1040.0 / 1.5);
-    }
-
-    #[test]
-    fn clamp_outer_position_moves_window_below_work_area_top() {
-        let work_area = PhysicalRect {
-            x: 0,
-            y: 40,
-            width: 1920,
-            height: 1040,
-        };
-        let outer_size = PhysicalWindowSize {
-            width: 1200,
-            height: 800,
-        };
-
-        let pos = clamp_outer_position_to_work_area(-100, -200, outer_size, work_area);
-
-        assert_eq!(pos.x, 0);
-        assert_eq!(pos.y, 40);
-    }
-
-    #[test]
-    fn clamp_outer_position_keeps_title_bar_visible_when_window_is_taller_than_work_area() {
-        let work_area = PhysicalRect {
-            x: 100,
-            y: 100,
-            width: 800,
-            height: 500,
-        };
-        let outer_size = PhysicalWindowSize {
-            width: 900,
-            height: 700,
-        };
-
-        let pos = clamp_outer_position_to_work_area(20, 20, outer_size, work_area);
-
-        assert_eq!(pos.x, 100);
-        assert_eq!(pos.y, 100);
     }
 
     #[test]
